% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADvariantExplorer.R
\name{ADvariantExplorer}
\alias{ADvariantExplorer}
\title{ADvariantExplorer}
\description{
A template for building documented, tested R6 classes
}
\examples{
  rt <- R6Template$new(id="abc")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{identifier for a class object}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{id}}{identifier for a class object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ADvariantExplorer$new()}}
\item \href{#method-getID}{\code{ADvariantExplorer$getID()}}
\item \href{#method-clone}{\code{ADvariantExplorer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ADvariantExplorer$new(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{character, an indentifier for this object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new instance of ADvariantExplorer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getID"></a>}}
\if{latex}{\out{\hypertarget{method-getID}{}}}
\subsection{Method \code{getID()}}{
accessor for the object's id field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ADvariantExplorer$getID()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the current value of the id member
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ADvariantExplorer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
