% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQTLCatalogFetcher.R
\name{EQTLCatalogFetcher}
\alias{EQTLCatalogFetcher}
\title{EQTLCatalogFetcher}
\description{
a small class to fetch eqtl regions by study id from the EBI eQTL catalog
}
\examples{
  studyIDs <- "GTEx_ge_brain_frontal_cortex"
  chrom <- "8"
  start <- 27610984
  end <- 27610987

  fetcher <- EQTLCatalogFetcher$new(studyIDs, chrom, start, end, simplify=TURE)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EQTLCatalogFetcher$new()}}
\item \href{#method-fetch}{\code{EQTLCatalogFetcher$fetch()}}
\item \href{#method-clone}{\code{EQTLCatalogFetcher$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EQTLCatalogFetcher$new(studyIDs, chrom, start, end, simplify = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{studyIDs}}{character vector, e.g. "GTEx_ge_brain_frontal_cortex"}

\item{\code{chrom}}{character, chromosome identifier, with or without leading "chr"}

\item{\code{start}}{numeric starting base pair of the region of interest}

\item{\code{end}}{numeric ending base pair of the region of interest}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new instance of EQTLCatalogFetcher
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fetch"></a>}}
\if{latex}{\out{\hypertarget{method-fetch}{}}}
\subsection{Method \code{fetch()}}{
obtain the requested slice of eQTLs from each study
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EQTLCatalogFetcher$fetch()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EQTLCatalogFetcher$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
